+++ CHANGELOG.md
@@ -3,6 +3,7 @@ We use [semantic versioning][semver]
 # Next version
 - [feature] added `dump-on-exit` option
 - [breaking change] added `mode` option (Must be set for Testwise Coverage mode. `http-server-port` or `test-env` alone is no longer sufficient)
+- [feature] The agent now optionally accepts test execution results via /test/end
 
 # 11.3.0
 - [breaking change] The convert tool now uses wildcard patterns for the class matching (was ant pattern before)